tests/wpt/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.interpolate.overlap2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.gradient.interpolate.overlap2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.interpolate.overlap2</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var g = ctx.createLinearGradient(0, 0, 100, 0);
var ps = [ 0, 1/10, 1/4, 1/3, 1/2, 3/4, 1 ];
for (var p = 0; p < ps.length; ++p)
{
        g.addColorStop(ps[p], '#0f0');
        for (var i = 0; i < 15; ++i)
                g.addColorStop(ps[p], '#f00');
        g.addColorStop(ps[p], '#0f0');
}
ctx.fillStyle = g;
ctx.fillRect(0, 0, 100, 50);
_assertPixel(offscreenCanvas, 1,25, 0,255,0,255, "1,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 30,25, 0,255,0,255, "30,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 40,25, 0,255,0,255, "40,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 60,25, 0,255,0,255, "60,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 80,25, 0,255,0,255, "80,25", "0,255,0,255");
t.done();

});
</script>
